+++ src/router/index.ts
@@ -1,15 +1,9 @@
 import { createRouter, createWebHashHistory, RouteRecordRaw } from "vue-router";
 import Home from "@/views/Home.vue";
-import About from "@/views/About.vue";
 
 const routes: Array<RouteRecordRaw> = [
   {
     path: "/",
-    name: "About",
-    component: About,
-  },
-  {
-    path: "/about",
     name: "Home",
     component: Home,
   },
